LayoutTests/http/tests/xmlhttprequest/access-control-response-with-expose-headers-expected.txt

CONSOLE MESSAGE: Refused to get unsafe header "X-TEST"
Test for bug 41210: Cross Origin XMLHttpRequest can not expose headers indicated in Access-Control-Expose-Headers HTTP Response Header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.getResponseHeader("X-FOO") is 'BAR'
PASS xhr.getResponseHeader("x-foo") is 'BAR'
PASS xhr.getResponseHeader("X-TEST") is null
PASS successfullyParsed is true

TEST COMPLETE
